Move status enum to Sidecar class
